Fixed function

+++ 14-binary_tree_balance.c
@@ -17,7 +17,7 @@ int binary_tree_balance(const binary_tree_t *tree)
 	left_height = binary_tree_height(tree->left);
 	right_height = binary_tree_height(tree->right);
 
-	return ((int)(left_height - right_height));
+	return (left_height - right_height);
 }
 /**
  * binary_tree_height - Computes the height of a binary tree
